+++ tests/testthat/test-api.R
@@ -17,6 +17,8 @@ context('api')
 
 
 # ---- classes, methods and predicates ------------------------------------------------------------
+
+
 test_that('elastic objects have the correct classes assigned to them', {
   # skip if on CRAN or Travis
   skip_on_travis()
@@ -131,6 +133,40 @@ test_that('sort objects generate the correct search API call', {
 })
 
 
+test_that('select_fields objects have the correct classes assigned to them', {
+  # arrange
+  select <- '{"includes": ["field1", "obj1.*"], "excludes": ["field2", "obj2.*"]}'
+
+  # act
+  es_source_filter <- select_fields(select)
+
+  # assert
+  expect_identical(class(es_source_filter), c("elastic_source_filter", "elastic_api", "elastic"))
+})
+
+
+test_that('select_fields objects will not accept invalid JSON', {
+  # arrange
+  bad_source_filter_json <- '{"includes": ["field1", "obj1.*"], "excludes": ["field2", "obj2.*"}'
+
+  # act & assert
+  expect_error(select_fields(bad_source_filter_json))
+})
+
+
+test_that('select_fields objects generate the correct search API call', {
+  # arrange
+  select <- '{"includes": ["field1", "obj1.*"], "excludes": ["field2"]}'
+
+  # act
+  es_source_filter <- select_fields(select)
+
+  # assert
+  expected_call <- '"_source": {"includes": ["field1", "obj1.*"], "excludes": ["field2"]}'
+  expect_identical(es_source_filter$api_call, expected_call)
+})
+
+
 test_that('aggs objects have the correct classes assigned to them', {
   # arrange
   avg_sepal_width_per_cat <- '{"avg_sepal_width_per_cat": {
@@ -174,7 +210,62 @@ test_that('aggs objects generate the correct search API call', {
 })
 
 
+test_that('list_indices objects have the correct classes assigned to them', {
+  # act
+  es_info <- list_indices()
+
+  # assert
+  expect_identical(class(es_info), c("elastic_info", "elastic_api", "elastic"))
+})
+
+
+test_that('list_fields objects have the correct classes assigned to them', {
+  # act
+  es_info <- list_fields()
+
+  # assert
+  expect_identical(class(es_info), c("elastic_info", "elastic_api", "elastic"))
+})
+
+
 # ---- operators ----------------------------------------------------------------------------------
+
+
+test_that('%info% list_indices() returns a list of all available indices', {
+  # skip if on CRAN or Travis
+  skip_on_travis()
+  skip_on_cran()
+
+  # arrange
+  load_test_data()
+
+  # act
+  info_results <- elastic("http://localhost:9200", "iris", "data") %info% list_indices()
+
+  # assert
+  expect_equal(info_results, "iris")
+  delete_test_data()
+})
+
+
+test_that('%info% list_fields() returns a list of all fields in an index', {
+  # skip if on CRAN or Travis
+  skip_on_travis()
+  skip_on_cran()
+
+  # arrange
+  load_test_data()
+
+  # act
+  info_results <- elastic("http://localhost:9200", "iris", "data") %info% list_fields()
+
+  # assert
+  expected_fields <- c("petal_length", "petal_width", "sepal_length", "sepal_width", "sort_key", "species")
+  expect_equal(info_results, expected_fields)
+  delete_test_data()
+})
+
+
 test_that('%index% correctly indexes a large (>10mb single chunk) data frame', {
   # skip if on CRAN or Travis
   skip_on_travis()
@@ -311,7 +402,7 @@ test_that('we can query using the %search% operator and return all documents', {
 })
 
 
-test_that('we can query using the %search% operator on a subset of all documents', {
+test_that('we can query using the %search% operator and return documents sorted', {
   # skip if on CRAN or Travis
   skip_on_travis()
   skip_on_cran()
@@ -330,6 +421,62 @@ test_that('we can query using the %search% operator on a subset of all documents
 
   # assert
   expect_equal(query_results_sorted, iris_data[1:10, ])
+  delete_test_data()
+})
+
+
+test_that('we can query using the %search% operator and return a subset of fields', {
+  # skip if on CRAN or Travis
+  skip_on_travis()
+  skip_on_cran()
+
+  # arrange
+  load_test_data()
+  fields <- c("sort_key", "sepal_length", "species")
+  everything <- '{"match_all": {}}'
+  source_filter_JSON <- '{"includes": ["sepal_length", "species", "sort_key"]}'
+  es_query <- query(everything)
+  es_source_filter <- select_fields(source_filter_JSON)
+
+  # act
+  query_results <-
+    elastic("http://localhost:9200", "iris", "data") %search% (es_query + es_source_filter)
+
+  query_results_sorted <- query_results[order(query_results["sort_key"]), fields]
+  rownames(query_results_sorted) <- query_results_sorted$sort_key
+
+  # assert
+  expect_equal(query_results_sorted, iris_data[, fields])
+  delete_test_data()
+})
+
+
+test_that('we can query using the %search% operator and return a sorted subset of fields', {
+  # skip if on CRAN or Travis
+  skip_on_travis()
+  skip_on_cran()
+
+  # arrange
+  load_test_data()
+  everything <- '{"match_all": {}}'
+  by_sepal_length <- '[{"sepal_length": {"order": "asc"}}, {"sort_key": {"order": "asc"}}]'
+  source_filter_JSON <- '{"includes": ["sepal_length", "species", "sort_key"]}'
+
+  es_query <- query(everything)
+  es_sort <- sort_on(by_sepal_length)
+  es_source_filter <- select_fields(source_filter_JSON)
+  es_filter_and_sort <- es_source_filter + es_sort
+
+  # act
+  query_results <-
+    elastic("http://localhost:9200", "iris", "data") %search% (es_query + es_filter_and_sort)
+
+  rownames(query_results) <- 1:150
+
+  # assert
+  iris_data_sorted <- iris_data[order(iris_data$sepal_length), colnames(query_results)]
+  rownames(iris_data_sorted) <- 1:150
+  expect_equal(query_results, iris_data_sorted)
   delete_test_data()
 })