+++ NEWS.md
@@ -1,3 +1,11 @@
+# elasticsearchr 0.3.0
+
+* Added support for source filtering with the `select_fields` function.
+* Added the `%info%` operator to be used for retreiving cluster and index information.
+* Added the `list_indices` function for retrieving a list of all available indices.
+* Added the `list_fields` function for retrieving a list of all available indices.
+
+
 # elasticsearchr 0.2.3
 
 * Refactored `valid_url` with the enhanced `valid_connection` function, based on HTTP requests to the cluster health API.